sl_dkls23
Module key_export

Exports a threshold key to a singleton one by consolidating all shares of other nodes.

Enums

KeyExportError

Distributed key generation errors

Constants

KEYSHARE_EXPORT_TAG

Tag of an enrypted keyshare message

Functions

combine_shares

Helper method to combine the secret shares into the private key You can use all the shares or the threshold number of shares to combine the private key

decrypt_share

Decrypt share encrypted by export_keyshare()

export_keyshare

Export keyshare.

message_receivers

Generate message receiver map.

receive_keyshares

Receive exported key shares, combine them and calculate private key.