sl_dkls23
Module setup

Setup messages. Each mpc protocol: Keygen, Sign, Quorum change, Key Export in order to bootstrap the nodes need to setup necessary information such as protocol id, participant information, cryptographic keys, and tailored per MPC protocol parameters.

Modules

finish

Setup for Finish PreSignature

key_export

Setup for Key export

keygen

Setup for DKG

quorum_change

Setup for Quorum Change

sign

Setup for DSG

Structs

AllOtherParties

An iterator for parties in range 0..total except me.

ArcSigner

A wrapper around an Arc<T> that implements Signer for the inner type.

NoSignature

A zero-sized type representing an empty signature. This type is used when no actual signature is needed, but the type system requires a signature type. It implements SignatureEncoding with an empty byte array representation.

NoSigningKey

A zero-sized type representing a signing key that produces no signatures. This type is used when no actual signing is needed, but the type system requires a signing key type. It implements Signer<NoSignature> and always returns NoSignature when signing.

NoVerifyingKey

A verifying key that always succeeds verification. This type is used when no actual verification is needed, but the type system requires a verifying key type. It’s typically used when secure transport is already in place and message authenticity doesn’t need to be verified. The inner Vec<u8> is used as an identity ID.

Constants

ABORT_MESSAGE_TAG

Tag of a broadcast message indicating that sender won’t participate in the protocol. The payload of the message contains error code.

SETUP_MESSAGE_TAG

Tag for all setup messages

Traits

FinalSignSetupMessage

A setup message for sign::finish()

KeyExportReceiverSetupMessage

A setup message for a reciever of exported key.

KeyExporterSetupMessage

A setup message for key export.

KeygenSetupMessage

A setup message for keygen::run()

PreSignSetupMessage

A setup message for sign::pre_signature()

ProtocolParticipant

Type that provides a protocol participant details.

QuorumChangeSetupMessage

A setup message for quorum_change::run()

SignSetupMessage

A setup message for sign::run()

SignatureEncoding

Support for decoding/encoding signatures as bytes.

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

Verifier

Verify the provided message bytestring using Self (e.g. a public key)